feat(webcodecs): align API with W3C WebCodecs specification
BREAKING CHANGE: This commit introduces several breaking changes to align
with the W3C WebCodecs specification.

## Type Changes

- ColorSpaceConversion enum values changed from 'Default'/'None' to
  'default'/'none' (lowercase per W3C spec)
- AudioData.sampleRate changed from u32 to f64 (W3C spec uses float)
- AudioEncoderConfig.sampleRate changed from u32 to f64
- AudioDecoderConfig.sampleRate changed from u32 to f64
- AlphaOption default changed from 'keep' to 'discard' (W3C spec default)

## VideoColorSpace Enums

Added proper enums instead of string types:
- VideoColorPrimaries: bt709, bt470bg, smpte170m, bt2020, smpte432
- VideoTransferCharacteristics: bt709, smpte170m, iec61966-2-1, linear, pq, hlg
- VideoMatrixCoefficients: rgb, bt709, bt470bg, smpte170m, bt2020-ncl

## Video Encoder Codec-Specific Configs (W3C Codec Registry)

- AvcBitstreamFormat enum (avc, annexb)
- AvcEncoderConfig with format option
- HevcBitstreamFormat enum (hevc, annexb)
- HevcEncoderConfig with format option
- VideoEncoderConfig now accepts avc/hevc codec-specific options
- VideoEncoderEncodeOptions now accepts codec-specific quantizer options:
  - AVC: quantizer 0-51
  - HEVC: quantizer 0-51
  - VP9: quantizer 0-63
  - AV1: quantizer 0-63

## Audio Encoder Codec-Specific Configs (W3C Codec Registry)

- OpusBitstreamFormat enum (opus, ogg)
- OpusSignal enum (auto, music, voice)
- OpusApplication enum (voip, audio, lowdelay)
- OpusEncoderConfig with full W3C spec fields (format, signal, application,
  frame_duration, complexity, packetlossperc, useinbandfec, usedtx)
- AacBitstreamFormat enum (aac, adts)
- AacEncoderConfig with format option
- FlacEncoderConfig with block_size and compress_level
- AudioEncoderConfig now accepts opus/aac/flac codec-specific options

## Metadata Additions

- SvcOutputMetadata struct with temporal_layer_id for SVC support
- EncodedVideoChunkMetadata now includes svc and alpha_side_data fields

+++ src/webcodecs/audio_decoder.rs
@@ -390,10 +390,10 @@ impl AudioDecoder {
       }
     };
 
-    // Configure decoder
+    // Configure decoder (cast f64 sample_rate to u32 for FFmpeg)
     let decoder_config = InternalAudioDecoderConfig {
       codec_id,
-      sample_rate: config.sample_rate,
+      sample_rate: config.sample_rate as u32,
       channels: config.number_of_channels,
       thread_count: 0, // Auto
       extradata: config.description.as_ref().map(|d| d.to_vec()),
